CENG493_Project/requirements.txt
faiss-cpu==1.9.0
sentence-transformers==3.4.1
transformers==4.57.6
torch==2.2.2
openai==1.63.2
rouge-score==0.1.2
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
httpx
rank_bm25>=0.2.2
ranx>=0.3.5
langchain-text-splitters>=0.3.0
scipy>=1.12.0
evaluate>=0.4.0
tqdm>=4.66.0
